requirements.prod.txt
# Core dependencies
aiogram>=3.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
asyncpg>=0.28.0
redis>=4.5.0

# Text processing and extraction
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
aiohttp>=3.8.0
trafilatura>=2.0.0

# Basic text analysis
langdetect>=1.0.9

# HTTP client for ML service
httpx>=0.24.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
